Add data export function to visualization panel
The changes mainly focus on implementing a data export function on the visualization panel and are also making some updates on the chart display method. Now, the displayChart function includes demographic data as a parameter, and an export function has been added to the DataInputPanel which allows users to export the displayed data as either an image or as a CSV file.